Do not assume resolved cache config in pre-join op
When cache service is creating its pre-join
operation, it creates a PreJoinCacheConfig
assuming that the CacheConfig has been resolved
(ie KV types and other user customizations
have been actually loaded). However it might
be the case that even though the CacheConfig is
registered in cache service, the cache has not
yet been touched, so user customizations are not
resolved yet.