Do not nullify containers upon shutdown (#16931)
The core reason of the NPE issue is late termination of MetricsRegistryImpl 
scheduled task. Before task termination, we shutdown MultiMapService and this 
causes task to find some service resources as null.

To easily fix the issue, this PR removes nullifying of containers in 
MultiMapService to align behavior with other services.